requirements.txt
python-telegram-bot==21.6
mistralai==1.10.0
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
apscheduler==3.10.4
streamlit>=1.35.0
python-dotenv==1.0.0
pytz==2024.1
beautifulsoup4==4.12.2
